pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-feedback"
version = "0.1.0"
description = "CLI tool to generate technical interview feedback reports from notes using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "reportlab>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pyinstaller>=6.0"]

[project.scripts]
feedback = "ai_feedback.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
